zoup/requirements.txt
Django~=4.1.3
python-dotenv
psycopg2
drf-yasg~=1.21.4
djangorestframework~=3.14.0
pyjwt~=2.6.0
gunicorn~=20.1